Code refactor for simplification (part 1)
