+++ v0.5/recommendation/python/main.py
@@ -21,10 +21,7 @@ import mlperf_loadgen as lg
 import numpy as np
 
 import dataset
-import imagenet
-import coco
-import criteoterabyte
-import dlrm_postprocess
+import criteo
 
 logging.basicConfig(level=logging.INFO)
 log = logging.getLogger("main")
@@ -36,33 +33,12 @@ MILLI_SEC = 1000
 
 # the datasets we support
 SUPPORTED_DATASETS = {
-    "imagenet":
-        (imagenet.Imagenet, dataset.pre_process_vgg, dataset.PostProcessCommon(offset=-1),
-         {"image_size": [224, 224, 3]}),
-    "imagenet_mobilenet":
-        (imagenet.Imagenet, dataset.pre_process_mobilenet, dataset.PostProcessArgMax(offset=-1),
-         {"image_size": [224, 224, 3]}),
-    "coco-300":
-        (coco.Coco, dataset.pre_process_coco_mobilenet, coco.PostProcessCoco(),
-         {"image_size": [300, 300, 3]}),
-    "coco-300-pt":
-        (coco.Coco, dataset.pre_process_coco_pt_mobilenet, coco.PostProcessCocoPt(False,0.3),
-         {"image_size": [300, 300, 3]}),
-    "coco-1200":
-        (coco.Coco, dataset.pre_process_coco_resnet34, coco.PostProcessCoco(),
-         {"image_size": [1200, 1200, 3]}),
-    "coco-1200-onnx":
-        (coco.Coco, dataset.pre_process_coco_resnet34, coco.PostProcessCocoOnnx(),
-         {"image_size": [1200, 1200, 3]}),
-    "coco-1200-pt":
-        (coco.Coco, dataset.pre_process_coco_resnet34, coco.PostProcessCocoPt(True,0.05),
-         {"image_size": [1200, 1200, 3],"use_label_map": True}),
-    "coco-1200-tf":
-        (coco.Coco, dataset.pre_process_coco_resnet34, coco.PostProcessCocoTf(),
-         {"image_size": [1200, 1200, 3],"use_label_map": False}),
-    "criteo-terabyte":
-        (criteoterabyte.CriteoTerabyte, dataset.pre_process_criteoterabyte_dlrm, dlrm_postprocess.DlrmPostProcess(),
-         {"max_ind_range": 40000, "sub_sample_rate": 0.0, "randomize": 'total',  "memory_map": True}),
+    "kaggle":
+        (criteo.Criteo, dataset.pre_process_criteo_dlrm, criteo.DlrmPostProcess(),
+         {"randomize": 'total',  "memory_map": True}),    
+    "terabyte":
+        (criteo.Criteo, dataset.pre_process_criteo_dlrm, criteo.DlrmPostProcess(),
+         {"max_ind_range": 40000000, "sub_sample_rate": 0.0, "randomize": 'total',  "memory_map": True}),
 }
 
 # pre-defined command line options so simplify things. They are used as defaults and can be
@@ -70,107 +46,28 @@ SUPPORTED_DATASETS = {
 
 SUPPORTED_PROFILES = {
     "defaults": {
-        "dataset": "imagenet",
-        "backend": "tensorflow",
-        "cache": 0,
-        "max-batchsize": 32,
-    },
-
-    # resnet
-    "resnet50-tf": {
-        "inputs": "input_tensor:0",
-        "outputs": "ArgMax:0",
-        "dataset": "imagenet",
-        "backend": "tensorflow",
-        "model-name": "resnet50",
-    },
-    "resnet50-onnxruntime": {
-        "dataset": "imagenet",
-        "outputs": "ArgMax:0",
-        "backend": "onnxruntime",
-        "model-name": "resnet50",
-    },
-
-    # mobilenet
-    "mobilenet-tf": {
-        "inputs": "input:0",
-        "outputs": "MobilenetV1/Predictions/Reshape_1:0",
-        "dataset": "imagenet_mobilenet",
-        "backend": "tensorflow",
-        "model-name": "mobilenet",
-    },
-    "mobilenet-onnxruntime": {
-        "dataset": "imagenet_mobilenet",
-        "outputs": "MobilenetV1/Predictions/Reshape_1:0",
-        "backend": "onnxruntime",
-        "model-name": "mobilenet",
-    },
-
-    # ssd-mobilenet
-    "ssd-mobilenet-tf": {
-        "inputs": "image_tensor:0",
-        "outputs": "num_detections:0,detection_boxes:0,detection_scores:0,detection_classes:0",
-        "dataset": "coco-300",
-        "backend": "tensorflow",
-        "model-name": "ssd-mobilenet",
-    },
-    "ssd-mobilenet-pytorch": {
-        "inputs": "image",
-        "outputs": "bboxes,labels,scores",
-        "dataset": "coco-300-pt",
-        "backend": "pytorch-native",
-        "model-name": "ssd-mobilenet",
-    },
-    "ssd-mobilenet-onnxruntime": {
-        "dataset": "coco-300",
-        "outputs": "num_detections:0,detection_boxes:0,detection_scores:0,detection_classes:0",
-        "backend": "onnxruntime",
-        "data-format": "NHWC",
-        "model-name": "ssd-mobilenet",
-    },
-
-    # ssd-resnet34
-    "ssd-resnet34-tf": {
-        "inputs": "image:0",
-        "outputs": "detection_bboxes:0,detection_classes:0,detection_scores:0",
-        "dataset": "coco-1200-tf",
-        "backend": "tensorflow",
-        "data-format": "NCHW",
-        "model-name": "ssd-resnet34",
-    },
-    "ssd-resnet34-pytorch": {
-        "inputs": "image",
-        "outputs": "bboxes,labels,scores",
-        "dataset": "coco-1200-pt",
-        "backend": "pytorch-native",
-        "model-name": "ssd-resnet34",
-    },
-    "ssd-resnet34-onnxruntime": {
-        "dataset": "coco-1200-onnx",
-        "inputs": "image",
-        "outputs": "bboxes,labels,scores",
-        "backend": "onnxruntime",
-        "data-format": "NCHW",
-        "max-batchsize": 1,
-        "model-name": "ssd-resnet34",
-    },
-    "ssd-resnet34-onnxruntime-tf": {
-        "dataset": "coco-1200-tf",
-        "inputs": "image:0",
-        "outputs": "detection_bboxes:0,detection_classes:0,detection_scores:0",
-        "backend": "onnxruntime",
-        "data-format": "NHWC",
-        "model-name": "ssd-resnet34",
-    },
-    "dlrm-pytorch": {
-        "dataset": "criteo-terabyte",
-        "inputs": "inputs are not used in dlrm",
+        "dataset": "terabyte",
+        "inputs": "continuous and categorical features",
         "outputs": "probability",
-        #"backend": "pytorch-native",
-        "backend": "pytorch-native-dlrm",
-        "data-format": "floats_and_offsets_indices",
-        "model-name": "dlrm",
-        "max-batchsize": 32,
+        "backend": "pytorch-native",
+        "model": "dlrm",
+        "max-batchsize": 2048,
+    },
+    "dlrm-kaggle-pytorch": {
+        "dataset": "kaggle",
+        "inputs": "continuous and categorical features",
+        "outputs": "probability",
+        "backend": "pytorch-native",
+        "model": "dlrm",
+        "max-batchsize": 128,
+    },
+    "dlrm-terabyte-pytorch": {
+        "dataset": "terabyte",
+        "inputs": "continuous and categorical features",
+        "outputs": "probability",
+        "backend": "pytorch-native",
+        "model": "dlrm",
+        "max-batchsize": 2048,
     },
 }
 
@@ -187,20 +84,18 @@ last_timeing = []
 def get_args():
     """Parse commandline."""
     parser = argparse.ArgumentParser()
+    parser.add_argument("--model", help="name of the mlperf model, ie. dlrm")
+    parser.add_argument("--model-path", required=True, help="path to the model file")
     parser.add_argument("--dataset", choices=SUPPORTED_DATASETS.keys(), help="dataset")
     parser.add_argument("--dataset-path", required=True, help="path to the dataset")
-    parser.add_argument("--dataset-list", help="path to the dataset list")
-    parser.add_argument("--data-format", choices=["NCHW", "NHWC"], help="data format")
     parser.add_argument("--profile", choices=SUPPORTED_PROFILES.keys(), help="standard profiles")
     parser.add_argument("--scenario", default="SingleStream",
                         help="mlperf benchmark scenario, one of " + str(list(SCENARIO_MAP.keys())))
     parser.add_argument("--max-batchsize", type=int, help="max batch size in a single inference")
-    parser.add_argument("--model", required=True, help="model file")
     parser.add_argument("--output", help="test results")
     parser.add_argument("--inputs", help="model inputs")
     parser.add_argument("--outputs", help="model outputs")
     parser.add_argument("--backend", help="runtime to use")
-    parser.add_argument("--model-name", help="name of the mlperf model, ie. resnet50")
     parser.add_argument("--threads", default=os.cpu_count(), type=int, help="threads")
     parser.add_argument("--qps", type=int, help="target qps")
     parser.add_argument("--cache", type=int, default=0, help="use cache")
@@ -242,6 +137,9 @@ def get_backend(backend):
     if backend == "tensorflow":
         from backend_tf import BackendTensorflow
         backend = BackendTensorflow()
+    elif backend == "tflite":
+        from backend_tflite import BackendTflite
+        backend = BackendTflite()
     elif backend == "onnxruntime":
         from backend_onnxruntime import BackendOnnxruntime
         backend = BackendOnnxruntime()
@@ -253,13 +151,10 @@ def get_backend(backend):
         backend = BackendPytorch()
     elif backend == "pytorch-native":
         from backend_pytorch_native import BackendPytorchNative
-        backend = BackendPytorchNative()
-    elif backend == "pytorch-native-dlrm":
-        from backend_pytorch_native_dlrm import BackendPytorchNativeDLRM
         # NOTE: pass model parameters here, the following options are available
         #
         # 1. Criteo Kaggle Display Advertisement Challenge Dataset (see ./bench/dlrm_s_criteo_kaggle.sh)
-        # backend = BackendPytorchNativeDLRM(
+        # backend = BackendPytorchNative(
         #    m_spa=16,
         #    ln_emb=np.array([1460,583,10131227,2202608,305,24,12517,633,3,93145,5683,8351593,3194,27,14992,5461306,10,5652,2173,4,7046547,18,15,286181,105,142572]),
         #    ln_bot=np.array([13,512,256,64,16]),
@@ -267,7 +162,7 @@ def get_backend(backend):
         #  )
         #
         # 2. Criteo Terabyte (see ./bench/dlrm_s_criteo_terabyte.sh [--sub-sample=0.875] --max-in-range=10000000)
-        # backend = BackendPytorchNativeDLRM(
+        # backend = BackendPytorchNative(
         #    m_spa=64,
         #    ln_emb=np.array([10000000,39060,17295,7424,20265,3,7122,1543,63,10000000,3067956,405282,10,2209,11938,155,4,976,14,10000000,10000000,10000000,590152,12973,108,36]),
         #    with sub-sampling 9980333,36084,17217,7378,20134,3,7112,1442,61, 9758201,1333352,313829,10,2208,11156,122,4,970,14, 9994222, 7267859, 9946608,415421,12420,101, 3
@@ -276,21 +171,18 @@ def get_backend(backend):
         # )
         #
         # 3. Criteo Terabyte MLPerf training (see ./bench/run_and_time.sh --max-in-range=40000000)
-        # backend = BackendPytorchNativeDLRM(
+        # backend = BackendPytorchNative(
         #    m_spa=128,
         #    ln_emb=np.array([40000000,39060,17295,7424,20265,3,7122,1543,63,40000000,3067956,405282,10,2209,11938,155,4,976,14,40000000,40000000,40000000,590152,12973,108,36]),
         #    ln_bot=([13,512,256,128]),
         #    ln_top=([479,1024,1024,512,256,1])
         # )
-        backend = BackendPytorchNativeDLRM(
+        backend = BackendPytorchNative(
             m_spa=16,
             ln_emb=np.array([1460,583,10131227,2202608,305,24,12517,633,3,93145,5683,8351593,3194,27,14992,5461306,10,5652,2173,4,7046547,18,15,286181,105,142572]),
             ln_bot=np.array([13,512,256,64,16]),
             ln_top=np.array([367,512,256,1])
         )
-    elif backend == "tflite":
-        from backend_tflite import BackendTflite
-        backend = BackendTflite()
     else:
         raise ValueError("unknown backend: " + backend)
     return backend
@@ -552,14 +444,11 @@ def add_results(final_results, name, result_dict, result_list, took, show_accura
 def main():
     global last_timeing
     args = get_args()
-
+    
     log.info(args)
 
     # find backend
     backend = get_backend(args.backend)
-
-    # override image format if given
-    image_format = args.data_format if args.data_format else backend.image_format()
 
     # --count applies to accuracy mode only and can be used to limit the number of images
     # for testing. For perf model we always limit count to 200.
@@ -572,14 +461,12 @@ def main():
     wanted_dataset, pre_proc, post_proc, kwargs = SUPPORTED_DATASETS[args.dataset]
 
     ds = wanted_dataset(data_path=args.dataset_path,
-                        image_list=args.dataset_list, # not needed - remove
                         name=args.dataset,
-                        image_format=image_format, # not needed - remove
-                        pre_process=pre_proc, # not needed - remove
-                        use_cache=args.cache,
+                        pre_process=pre_proc,  # currently an identity function
+                        use_cache=args.cache,  # currently not used
                         count=count, **kwargs)
     # load model to backend
-    model = backend.load(args.model, inputs=args.inputs, outputs=args.outputs)
+    model = backend.load(args.model_path, inputs=args.inputs, outputs=args.outputs)
     final_results = {
         "runtime": model.name(),
         "version": model.version(),
@@ -604,42 +491,24 @@ def main():
     # warmup
     ds.load_query_samples([0])
 
-    if 'dlrm' in args.backend:
-        for _ in range(5):
-            batch_dense_X, batch_lS_o, batch_lS_i, batch_T = ds.get_samples([0])
-            _ = backend.predict(batch_dense_X, batch_lS_o, batch_lS_i)
-    else:
-        for _ in range(5):
-            img, _ = ds.get_samples([0])
-            _ = backend.predict({backend.inputs[0]: img})
+    for _ in range(5):
+        batch_dense_X, batch_lS_o, batch_lS_i, batch_T = ds.get_samples([0])
+        _ = backend.predict(batch_dense_X, batch_lS_o, batch_lS_i)
 
     ds.unload_query_samples(None)
 
-    if 'dlrm' in args.backend:
-        scenario = SCENARIO_MAP[args.scenario]
-        runner_map = {
-            lg.TestScenario.SingleStream: RunnerBase,
-            lg.TestScenario.MultiStream: QueueRunnerDlrm,
-            lg.TestScenario.Server: QueueRunnerDlrm,
-            lg.TestScenario.Offline: QueueRunnerDlrm
-            }
-    else:
-        scenario = SCENARIO_MAP[args.scenario]
-        runner_map = {
-            lg.TestScenario.SingleStream: RunnerBase,
-            lg.TestScenario.MultiStream: QueueRunner,
-            lg.TestScenario.Server: QueueRunner,
-            lg.TestScenario.Offline: QueueRunner
-            }
+    scenario = SCENARIO_MAP[args.scenario]
+    runner_map = {
+        lg.TestScenario.SingleStream: RunnerBase,
+        lg.TestScenario.MultiStream: QueueRunnerDlrm,
+        lg.TestScenario.Server: QueueRunnerDlrm,
+        lg.TestScenario.Offline: QueueRunnerDlrm
+    }
 
     runner = runner_map[scenario](model, ds, args.threads, post_proc=post_proc, max_batchsize=args.max_batchsize)
 
     def issue_queries(query_samples):
-
-        if 'dlrm' in args.backend:
-            runner.enqueue_dlrm(query_samples)
-        else:
-            runner.enqueue(query_samples)
+        runner.enqueue_dlrm(query_samples)
 
     def flush_queries():
         pass
@@ -650,7 +519,7 @@ def main():
         last_timeing = [t / NANO_SEC for t in latencies_ns]
 
     settings = lg.TestSettings()
-    settings.FromConfig(config, args.model_name, args.scenario)
+    settings.FromConfig(config, args.model, args.scenario)
     settings.scenario = scenario
     settings.mode = lg.TestMode.PerformanceOnly
     if args.accuracy:
