+++ v0.5/recommendation/run_local.sh
@@ -3,11 +3,11 @@
 source ./run_common.sh
 
 common_opt="--config ../mlperf.conf"
-dataset="--dataset-path $DATA_DIR"
 OUTPUT_DIR=`pwd`/output/$name
 if [ ! -d $OUTPUT_DIR ]; then
     mkdir -p $OUTPUT_DIR
 fi
 
-python python/main.py --profile $profile $common_opt --model $model_path $dataset \
-    --output $OUTPUT_DIR $EXTRA_OPS $@
+python python/main.py --profile $profile $common_opt --model $model --model-path $model_path \
+       --dataset $dataset --dataset-path $DATA_DIR \
+       --output $OUTPUT_DIR $EXTRA_OPS $@
